+++ lib/nylas/token_info.rb
@@ -18,5 +18,3 @@ module Nylas
     end
   end
 end
-
-