+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### Unreleased
+* Add job status support
+* Enable account deletion
+
 ### 5.6.1 / 2021-12-13
 * Enabled support for adding metadata to a `NewMessage`/`Draft`
 * Fix bug where updating an Event results in an API error