+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ### Unreleased
+* Add `Event` to ICS support
 * Added support for getting access token information
 * Improved support for Application Details